fix: upgrade volarjs/services to from v0.0.4 to v0.0.5
The published v0.0.4 (f5d4949) of volar-service-typescript causes error
because it tries to access ctx.typescript.languageService; in fact it
was `undefined`. After f5d4949, the v0.0.4 continues but no new version
of it was published on npmjs. v0.0.5 has just been released, in which
the above problem was fixed. Some API changes were compensated for.